Restructure camera build mode
