+++ sdk/src/cameras/ad-fxtof1-ebz/camera_fxtof1.h
@@ -45,8 +45,8 @@ class CameraFxTof1 : public aditof::Camera {
   public:
     CameraFxTof1(
         std::shared_ptr<aditof::DepthSensorInterface> depthSensor,
-        std::shared_ptr<aditof::StorageInterface> eeprom,
-        std::shared_ptr<aditof::TemperatureSensorInterface> temperatureSensor); 
+        std::vector<std::shared_ptr<aditof::StorageInterface>> &eeprom,
+        std::vector<std::shared_ptr<aditof::TemperatureSensorInterface>> &tSensor); 
     ~CameraFxTof1();
 
   public: // implements Camera
