+++ sdk/src/cameras/ad-fxtof1-ebz/jetson/target_definitions.h
@@ -33,12 +33,10 @@
 #define TARGET_DEFINITIONS_H
 
 static const char *EEPROM_NAME = "24c1024";
-static const char *EEPROM_REPLACEMENT_PATH = "";
+static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/6-0056/eeprom";
 
 static const char *TEMP_SENSOR_DEV_PATH = "/sys/class/hwmon/hwmon0/temp1_input";
-static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/1-0056/eeprom";
-static const char *CAPTURE_DEVICE_NAME = "rkisp10_mainpath";
 
-static const int NUM_VIDEO_DEVS = 1;
+static const char *CAPTURE_DEVICE_NAME = "vi-output, addi9036 6-0064";
 
 #endif // TARGET_DEFINITIONS_H
