+++ sdk/src/cameras/ad-fxtof1-ebz/raspberrypi/target_definitions.h
@@ -33,17 +33,10 @@
 #define TARGET_DEFINITIONS_H
 
 static const char *EEPROM_NAME = "24c1024";
-static const char *TEMP_SENSOR_DEV_PATH = "/dev/i2c-1";
-static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/1-0056/eeprom";
+static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/0-0056/eeprom";
 
-static const char *EEPROM_REPLACEMENT_PATH =
-    "/home/analog/workspace/github/aditof_sdk/chicony_firmware.bin";
+static const char *TEMP_SENSOR_DEV_PATH = "/sys/class/hwmon/hwmon2/temp1_input";
 
-static const char *TEMP_SENSOR_REPLACEMENT_DEV_PATH = "";
-
-static const char *CAPTURE_DEVICE_NAME =
-    "vi-output, addi9036 1-0064|vi-output, addi9036 7-0064";
-
-static const int NUM_VIDEO_DEVS = 2;
+static const char *CAPTURE_DEVICE_NAME = "unicam";
 
 #endif // TARGET_DEFINITIONS_H
