+++ sdk/src/cameras/ad-fxtof1-ebz/xavier_nx/target_definitions.h
@@ -29,31 +29,15 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "connections/target/target_sensor_enumerator.h"
-#include "target_definitions.h"
+#ifndef TARGET_DEFINITIONS_H
+#define TARGET_DEFINITIONS_H
 
-#include <dirent.h>
-#include <glog/logging.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <string>
-#include <sys/stat.h>
-#include <unistd.h>
+static const char *EEPROM_NAME = "24c1024";
+static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/9-0056/eeprom";
 
-using namespace aditof;
+static const char *CAPTURE_DEVICE_NAME = "vi-output, addi9036 6-0064";
 
-Status TargetSensorEnumerator::searchSensors() {
+static const char *CAPTURE_DEVICE_NAME =
+    "vi-output, addi9036 9-0064|vi-output, addi9036 10-0064";
 
-    LOG(INFO) << "Looking for devices on the target: Xavier";
-
-    // TO DO: Don't guess the device, find a way to identify it so we are sure
-    // we've got the right device and is compatible with the SDK
-    SensorInfo sInfo;
-    sInfo.sensorType = SensorType::SENSOR_ADDI9036;
-    sInfo.driverPath = "/dev/video0|/dev/video1";
-    sInfo.subDevPath = "/dev/v4l-subdev0|/dev/v4l-subdev2";
-    m_sensorsInfo.emplace_back(sInfo);
-
-    return Status::OK;
-}
+#endif // TARGET_DEFINITIONS_H
