+++ sdk/src/cameras/chicony-006/raspberrypi/target_definitions.h
@@ -29,42 +29,14 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef SENSOR_DEFINITIONS_H
-#define SENSOR_DEFINITIONS_H
+#ifndef TARGET_DEFINITIONS_H
+#define TARGET_DEFINITIONS_H
 
-#include <string>
+static const char *EEPROM_NAME = "24c1024";
 
-/**
- * @brief Namespace aditof
- */
-namespace aditof {
+static const char *EEPROM_DEV_PATH = "/sys/bus/i2c/devices/0-0056/eeprom";
+static const char *TEMP_SENSOR_DEV_PATH = "/sys/class/hwmon/hwmon2/temp1_input";
 
-/**
- * @enum SensorType
- * @brief Provides the types of sensor assosiated with the device
- */
-enum class SensorType {
-    SENSOR_ADDI9036, //!< ADDI9036 CCD sensor
-    SENSOR_ADT7410,  //!< ADT7410 temperature sensor
-    SENSOR_TMP10X,   //!< TMP101 / TMP102 temperature sensor
-};
+static const char *CAPTURE_DEVICE_NAME = "unicam";
 
-/**
- * @struct SensorDetails
- * @brief Provides details about the device
- */
-struct SensorDetails {
-    /**
-     * @brief The type of sensor
-     */
-    SensorType sensorType;
-
-    /**
-     * @brief The sensor's name
-     */
-    std::string sensorName;
-};
-
-} // namespace aditof
-
-#endif // SENSOR_DEFINITIONS_H
+#endif // TARGET_DEFINITIONS_H
